packages/neo4jdb/requirements.txt
shapely
neomodel
pandas
python-dotenv
